requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
scikit-learn==1.5.2
sentence-transformers==3.1.1
faiss-cpu==1.13.2
numpy
PyMuPDF
requests